fix: move to core package all the utility functions
